conf: implement dirty interface for tests
Conf will be rewritten in the future. For now, simple interface is
enough.

Conf is a subsystem which combines config files, command-line options
and run-time objects provided by modules. It provides key-value storage
interface.
Module specific keys will have a prefix with the module name. For
exmaple, "xmpp.jid". Prefixes are NOT treated in special way, but create
unique strings.